+++ django_be/requirements.txt
@@ -38,6 +38,7 @@ mccabe==0.7.0
 mmh3==4.0.1
 mypy-extensions==1.0.0
 nest-asyncio==1.5.6
+openai==1.11.1
 packaging==23.1
 parso==0.8.3
 pathspec==0.11.1
@@ -57,6 +58,7 @@ pyflakes==3.0.1
 Pygments==2.15.1
 pyOpenSSL==23.3.0
 pyotp==2.9.0
+PyPDF2==3.0.1
 pypng==0.20220715.0
 PySocks==1.7.1
 python-dateutil==2.8.2
@@ -79,3 +81,5 @@ wcwidth==0.2.6
 Werkzeug==2.3.6
 pymongo==4.6.1
 retrying==1.3.4
+docx2txt==0.8
+nltk==3.8.1
